backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
requests>=2.28.0
python-multipart>=0.0.5
jinja2>=3.0.0
aiofiles>=23.0.0
python-dotenv>=0.19.0
httpx>=0.24.0
